Add styles to components and refactor code
