+++ public/notes-content.json
@@ -5,7 +5,7 @@
     "time": "2023-04-01",
     "category": "Task",
     "content": "Tomatoes, bread",
-    "archived": true
+    "archived": false
   },
   {
     "id": 1,
