+++ src/components/Layout/Layout.css
@@ -1,0 +1,5 @@
+.main {
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 20px;
+}
