+++ src/components/Modal/Modal.css
@@ -1,32 +1,46 @@
 .modal {
   position: fixed;
-  top: 0;
+  z-index: 1;
   left: 0;
+  top: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  background-color: rgba(15, 13, 13, 0.628);
 }
 
 .modal-content {
   background-color: #fff;
-  padding: 20px;
-  border-radius: 5px;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.2);
+  margin: 15% auto;
+  padding: 40px;
+  width: 400px;
+  border-radius: 8px;
 }
 
-/* .close-btn {
-  margin-top: 10px;
-  padding: 8px 16px;
-  background-color: #333;
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
+.modal-content input,
+.modal-content select {
+  display: block;
+  width: 100%;
+  padding: 10px;
+  margin-bottom: 20px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  outline: none;
 }
 
-.close-btn:hover {
-  background-color: #555;
-} */
+.modal-content input:focus {
+  border-color: blue;
+}
+
+.modal-content input,
+.modal-content select:not(:last-child) {
+  margin-top: 6px;
+}
+
+.modal-content select {
+  margin-top: 6px;
+}
+
+.modal-content label {
+  font-weight: bold;
+  font-size: 14px;
+}
