--- src/components/Table/Table.css
@@ -1,5 +1,0 @@
-.container {
-  max-width: 100%;
-  padding: 20px;
-  margin: 0 auto;
-}
