+++ src/index.css
@@ -17,3 +17,59 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
+
+table {
+  width: 1000px;
+  border-spacing: separate;
+  border-spacing: 1px;
+  color: #656565;
+  margin-bottom: 20px;
+}
+
+th,
+td {
+  background-color: #dde0f86e;
+  padding: 8px;
+}
+
+th {
+  background-color: #9b9b9b;
+  color: #fff;
+  text-align: left;
+}
+
+tr:hover {
+  background-color: #ddd;
+}
+
+button {
+  background-color: #9b9b9b;
+  border-radius: 4px;
+  color: #fff;
+  font-weight: bold;
+  padding: 10px 14px;
+  border: none;
+  cursor: pointer;
+}
+
+button:hover {
+  background-color: #ddd;
+}
+
+button:not(:last-child) {
+  margin-right: 4px;
+}
+
+#root {
+  max-width: 1000px;
+  margin: 0 auto;
+}
+
+.buttons-container {
+  display: flex;
+  justify-content: center;
+}
+
+.button-add {
+  margin-bottom: 20px;
+}
