+++ src/pages/HomePage.css
@@ -1,0 +1,4 @@
+.container {
+  padding: 20px 0;
+  margin: 0 auto;
+}
