+++ src/utils/generateRandomId.ts
@@ -1,0 +1,5 @@
+const generateRandomId = (): number => {
+  return new Date().getTime();
+};
+
+export default generateRandomId;
