--- src/utils/randomIdGenerator.ts
@@ -1,3 +1,0 @@
-export function generateRandomId(): number {
-  return new Date().getTime();
-}
